search frontend: mint query.ts for query helper functions
