+++ client/shared/src/util/url.ts
@@ -4,8 +4,8 @@ import { Position, Range, Selection } from '@sourcegraph/extension-api-types'
 import { WorkspaceRootWithMetadata } from '../api/extension/extensionHostApi'
 import { SearchPatternType } from '../graphql-operations'
 import { discreteValueAliases } from '../search/query/filters'
+import { findFilter, FilterKind } from '../search/query/query'
 import { appendContextFilter } from '../search/query/transformer'
-import { findFilter, FilterKind } from '../search/query/validate'
 
 import { replaceRange } from './strings'
